pyproject.toml
[tool.poetry]
name = "final-project"
version = "0.1.0"
description = ""
authors = ["hj <charmsxda402@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2.16"
djangorestframework = "^3.15.2"
dj-rest-auth = "^7.0.0"
django-allauth = "^65.2.0"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
openai = "^1.55.0"
langchain = "^0.3.7"
langchain-community = "^0.3.7"
langchain-openai = "^0.2.9"
pgvector = "^0.3.6"
sentence-transformers = "^3.3.1"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
